+++ firestore.rules
@@ -38,9 +38,8 @@ service cloud.firestore {
       request.resource.data.feedItemId is string &&
       request.resource.data.feedItemId == feedItemId &&
       request.resource.data.feedItemType in ['ARTICLE', 'VIDEO', 'WEBSITE', 'TWEET', 'XKCD', 'YOUTUBE'] &&
-      request.resource.data.feedSource.keys().hasAll(['type']) &&
-      request.resource.data.feedSource.type in ['RSS', 'YOUTUBE_CHANNEL', 'INTERVAL', 'PWA', 'EXTENSION', 'POCKET_EXPORT'] &&
-      request.resource.data.feedSource.userFeedSubscriptionId is string &&
+      request.resource.data.feedSource.keys().hasAll(['feedSourceType']) &&
+      request.resource.data.feedSource.feedSourceType in ['RSS', 'YOUTUBE_CHANNEL', 'INTERVAL', 'PWA', 'EXTENSION', 'POCKET_EXPORT'] &&
       request.resource.data.url is string &&
       request.resource.data.title is string &&
       request.resource.data.outgoingLinks is list &&
