+++ packages/pwa/src/screens/ImportScreen.tsx
@@ -4,6 +4,7 @@ import {useState} from 'react';
 import {asyncTry, prefixError} from '@shared/lib/errorUtils.shared';
 import {pluralizeWithCount} from '@shared/lib/utils.shared';
 
+import {POCKET_EXPORT_FEED_SOURCE} from '@shared/types/feedItems.types';
 import {
   ExternalMigrationItemStatus,
   NOT_STARTED_EXTERNAL_MIGRATION_ITEM_STATE,
@@ -11,7 +12,6 @@ import {
 } from '@shared/types/migration.types';
 import type {ExternalMigrationItemState} from '@shared/types/migration.types';
 import type {PocketImportItem} from '@shared/types/pocket.types';
-import {POCKET_EXPORT_MINI_USER_FEED_SUBSCRIPTION} from '@shared/types/userFeedSubscriptions.types';
 import type {Task} from '@shared/types/utils.types';
 
 import {useFeedItemsService} from '@sharedClient/services/feedItems.client';
@@ -100,10 +100,10 @@ export const ImportScreen: React.FC = () => {
   const handleImportItem = async (item: PocketImportItem, key: string): Promise<void> => {
     setImportStatus(key, PROCESSING_EXTERNAL_MIGRATION_ITEM_STATE);
 
-    const createResult = await feedItemsService.createFeedItem(
-      POCKET_EXPORT_MINI_USER_FEED_SUBSCRIPTION,
-      {url: item.url, title: item.title}
-    );
+    const createResult = await feedItemsService.createFeedItem(POCKET_EXPORT_FEED_SOURCE, {
+      url: item.url,
+      title: item.title,
+    });
 
     if (!createResult.success) {
       setImportStatus(key, {
