+++ packages/shared/src/lib/feedItems.shared.ts
@@ -15,14 +15,14 @@ import {
   makeNewFeedItemImportState,
   TriageStatus,
 } from '@shared/types/feedItems.types';
-import type {FeedItem, FeedItemAction, FeedSource} from '@shared/types/feedItems.types';
+import type {FeedItem, FeedItemAction} from '@shared/types/feedItems.types';
+import type {FeedSource} from '@shared/types/feedSources.types';
 import {IconName} from '@shared/types/icons.types';
 import type {Result} from '@shared/types/results.types';
 import {KeyboardShortcutId} from '@shared/types/shortcuts.types';
 import {SystemTagId} from '@shared/types/tags.types';
 import type {TagId} from '@shared/types/tags.types';
 import type {
-  MiniUserFeedSubscription,
   UserFeedSubscription,
   UserFeedSubscriptionId,
 } from '@shared/types/userFeedSubscriptions.types';
