+++ packages/shared/src/lib/userFeedSubscriptions.shared.ts
@@ -2,7 +2,7 @@ import {IMMEDIATE_DELIVERY_SCHEDULE} from '@shared/lib/deliverySchedules.shared'
 import {makeUuid} from '@shared/lib/utils.shared';
 
 import type {AccountId} from '@shared/types/accounts.types';
-import {FeedSourceType} from '@shared/types/feedItems.types';
+import {FeedSourceType} from '@shared/types/feedSources.types';
 import type {
   IntervalUserFeedSubscription,
   RssUserFeedSubscription,
