+++ packages/shared/src/types/feedItems.types.ts
@@ -4,15 +4,12 @@ import {makeUuid} from '@shared/lib/utils.shared';
 
 import type {AccountId} from '@shared/types/accounts.types';
 import {AccountIdSchema} from '@shared/types/accounts.types';
+import type {FeedSource} from '@shared/types/feedSources.types';
 import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
 import type {IconName} from '@shared/types/icons.types';
 import type {KeyboardShortcutId} from '@shared/types/shortcuts.types';
 import type {TagId} from '@shared/types/tags.types';
-import type {UserFeedSubscriptionId} from '@shared/types/userFeedSubscriptions.types';
-import {UserFeedSubscriptionIdSchema} from '@shared/types/userFeedSubscriptions.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
-import {YouTubeChannelIdSchema} from '@shared/types/youtube.types';
-import type {YouTubeChannelId} from '@shared/types/youtube.types';
 
 /**
  * Strongly-typed type for a {@link FeedItem}'s unique identifier. Prefer this over plain strings.
@@ -39,78 +36,6 @@ export enum FeedItemType {
   Xkcd = 'XKCD',
   YouTube = 'YOUTUBE',
 }
-
-/**
- * The origin of the feed item. Where the feed item came from.
- */
-export enum FeedSourceType {
-  /** RSS feeds. */
-  RSS = 'RSS',
-  /** YouTube channels. */
-  YouTubeChannel = 'YOUTUBE_CHANNEL',
-  /** Dummy feeds that automatically generate items at a fixed interval. */
-  Interval = 'INTERVAL',
-  /** Feeds that are added from the PWA. */
-  PWA = 'PWA',
-  /** Feeds that are added from the web extension. */
-  Extension = 'EXTENSION',
-  /** Feeds that are added from a Pocket export. */
-  PocketExport = 'POCKET_EXPORT',
-}
-
-interface BaseFeedSource {
-  readonly feedSourceType: FeedSourceType;
-}
-
-export interface RssFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.RSS;
-  readonly userFeedSubscriptionId: UserFeedSubscriptionId;
-  readonly url: string;
-  readonly title: string;
-}
-
-export interface YouTubeChannelFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.YouTubeChannel;
-  readonly userFeedSubscriptionId: UserFeedSubscriptionId;
-  readonly channelId: YouTubeChannelId;
-}
-
-export interface IntervalFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.Interval;
-  readonly userFeedSubscriptionId: UserFeedSubscriptionId;
-}
-
-export interface PwaFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.PWA;
-}
-
-export const PWA_FEED_SOURCE: PwaFeedSource = {
-  feedSourceType: FeedSourceType.PWA,
-};
-
-export interface ExtensionFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.Extension;
-}
-
-export const EXTENSION_FEED_SOURCE: ExtensionFeedSource = {
-  feedSourceType: FeedSourceType.Extension,
-};
-
-export interface PocketExportFeedSource extends BaseFeedSource {
-  readonly feedSourceType: FeedSourceType.PocketExport;
-}
-
-export const POCKET_EXPORT_FEED_SOURCE: PocketExportFeedSource = {
-  feedSourceType: FeedSourceType.PocketExport,
-};
-
-export type FeedSource =
-  | RssFeedSource
-  | YouTubeChannelFeedSource
-  | IntervalFeedSource
-  | PwaFeedSource
-  | ExtensionFeedSource
-  | PocketExportFeedSource;
 
 export enum TriageStatus {
   Untriaged = 'UNTRIAGED',
@@ -255,49 +180,6 @@ const FeedItemImportStateFromStorageSchema = z.discriminatedUnion('status', [
 ]);
 
 export type FeedItemImportStateFromStorage = z.infer<typeof FeedItemImportStateFromStorageSchema>;
-
-const BaseFeedSourceSchema = z.object({
-  feedSourceType: z.nativeEnum(FeedSourceType),
-});
-
-export const RssFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.RSS),
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-  url: z.string().url(),
-  title: z.string(),
-});
-
-export const YouTubeChannelFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.YouTubeChannel),
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-  channelId: YouTubeChannelIdSchema,
-});
-
-export const IntervalFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.Interval),
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-});
-
-const PwaFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.PWA),
-});
-
-const ExtensionFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.Extension),
-});
-
-const PocketExportFeedSourceSchema = BaseFeedSourceSchema.extend({
-  feedSourceType: z.literal(FeedSourceType.PocketExport),
-});
-
-export const FeedSourceSchema = z.discriminatedUnion('feedSourceType', [
-  RssFeedSourceSchema,
-  YouTubeChannelFeedSourceSchema,
-  IntervalFeedSourceSchema,
-  PwaFeedSourceSchema,
-  ExtensionFeedSourceSchema,
-  PocketExportFeedSourceSchema,
-]);
 
 /**
  * Zod schema for a {@link FeedItem} persisted to Firestore.
