+++ packages/shared/src/types/userFeedSubscriptions.types.ts
@@ -4,7 +4,7 @@ import type {AccountId} from '@shared/types/accounts.types';
 import {AccountIdSchema} from '@shared/types/accounts.types';
 import type {DeliverySchedule} from '@shared/types/deliverySchedules.types';
 import {DeliveryScheduleFromStorageSchema} from '@shared/types/deliverySchedules.types';
-import {FeedSourceType} from '@shared/types/feedItems.types';
+import {FeedSourceType} from '@shared/types/feedSources.types';
 import {FirestoreTimestampSchema} from '@shared/types/firebase.types';
 import type {BaseStoreItem} from '@shared/types/utils.types';
 import {YouTubeChannelIdSchema, type YouTubeChannelId} from '@shared/types/youtube.types';
@@ -116,121 +116,3 @@ export const UserFeedSubscriptionFromStorageSchema = z.union([
  * Type for a {@link UserFeedSubscription} persisted to Firestore.
  */
 export type UserFeedSubscriptionFromStorage = z.infer<typeof UserFeedSubscriptionFromStorageSchema>;
-
-////////////////////////////////
-//  MiniUserFeedSubscription  //
-////////////////////////////////
-export interface PWAMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.PWA;
-}
-
-export const PWA_MINI_USER_FEED_SUBSCRIPTION: PWAMiniUserFeedSubscription = {
-  feedSourceType: FeedSourceType.PWA,
-};
-
-export interface ExtensionMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.Extension;
-}
-
-export const EXTENSION_MINI_USER_FEED_SUBSCRIPTION: ExtensionMiniUserFeedSubscription = {
-  feedSourceType: FeedSourceType.Extension,
-};
-
-export interface PocketExportMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.PocketExport;
-}
-
-export const POCKET_EXPORT_MINI_USER_FEED_SUBSCRIPTION: PocketExportMiniUserFeedSubscription = {
-  feedSourceType: FeedSourceType.PocketExport,
-};
-
-type BasePersistedMiniUserFeedSubscription = Pick<UserFeedSubscription, 'userFeedSubscriptionId'>;
-
-export interface RssMiniUserFeedSubscription extends BasePersistedMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.RSS;
-  readonly url: string;
-  readonly title: string;
-}
-
-export interface YouTubeChannelMiniUserFeedSubscription
-  extends BasePersistedMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.YouTubeChannel;
-  readonly channelId: YouTubeChannelId;
-}
-
-export interface IntervalMiniUserFeedSubscription extends BasePersistedMiniUserFeedSubscription {
-  readonly feedSourceType: FeedSourceType.Interval;
-}
-
-export type MiniUserFeedSubscription =
-  | PWAMiniUserFeedSubscription
-  | ExtensionMiniUserFeedSubscription
-  | PocketExportMiniUserFeedSubscription
-  | RssMiniUserFeedSubscription
-  | YouTubeChannelMiniUserFeedSubscription
-  | IntervalMiniUserFeedSubscription;
-
-const PWAMiniUserFeedSubscriptionSchema = z.object({
-  feedSourceType: z.literal(FeedSourceType.PWA),
-});
-
-const ExtensionMiniUserFeedSubscriptionSchema = z.object({
-  feedSourceType: z.literal(FeedSourceType.Extension),
-});
-
-const PocketExportMiniUserFeedSubscriptionSchema = z.object({
-  feedSourceType: z.literal(FeedSourceType.PocketExport),
-});
-
-const BasePersistedMiniUserFeedSubscriptionSchema = z.object({
-  feedSourceType: z.union([
-    z.literal(FeedSourceType.RSS),
-    z.literal(FeedSourceType.YouTubeChannel),
-    z.literal(FeedSourceType.Interval),
-  ]),
-  userFeedSubscriptionId: UserFeedSubscriptionIdSchema,
-});
-
-export const RssMiniUserFeedSubscriptionSchema = BasePersistedMiniUserFeedSubscriptionSchema.extend(
-  {
-    feedSourceType: z.literal(FeedSourceType.RSS),
-    url: z.string(),
-    title: z.string(),
-  }
-);
-
-export const YouTubeChannelMiniUserFeedSubscriptionSchema =
-  BasePersistedMiniUserFeedSubscriptionSchema.extend({
-    feedSourceType: z.literal(FeedSourceType.YouTubeChannel),
-    channelId: YouTubeChannelIdSchema,
-  });
-
-export const IntervalMiniUserFeedSubscriptionSchema =
-  BasePersistedMiniUserFeedSubscriptionSchema.extend({
-    feedSourceType: z.literal(FeedSourceType.Interval),
-  });
-
-/**
- * Zod schema for a {@link MiniUserFeedSubscription} persisted to Firestore.
- */
-export const MiniUserFeedSubscriptionFromStorageSchema = z.union([
-  PWAMiniUserFeedSubscriptionSchema,
-  ExtensionMiniUserFeedSubscriptionSchema,
-  PocketExportMiniUserFeedSubscriptionSchema,
-  RssMiniUserFeedSubscriptionSchema,
-  YouTubeChannelMiniUserFeedSubscriptionSchema,
-  IntervalMiniUserFeedSubscriptionSchema,
-]);
-
-export const PersistedMiniUserFeedSubscriptionFromStorageSchema = z.union([
-  RssMiniUserFeedSubscriptionSchema,
-  YouTubeChannelMiniUserFeedSubscriptionSchema,
-  IntervalMiniUserFeedSubscriptionSchema,
-]);
-
-/**
- * Type for a {@link MiniUserFeedSubscription} persisted to Firestore.
- */
-export type MiniUserFeedSubscriptionFromStorage = z.infer<
-  typeof MiniUserFeedSubscriptionFromStorageSchema
->;
