+++ packages/sharedClient/src/services/feedItems.client.ts
@@ -29,8 +29,9 @@ import {parseFeedItem, parseFeedItemId, toStorageFeedItem} from '@shared/parsers
 import type {AccountId, AuthStateChangedUnsubscribe} from '@shared/types/accounts.types';
 import {AsyncStatus} from '@shared/types/asyncState.types';
 import type {AsyncState} from '@shared/types/asyncState.types';
-import type {FeedItem, FeedItemId, FeedSource, XkcdFeedItem} from '@shared/types/feedItems.types';
-import {FeedSourceType} from '@shared/types/feedItems.types';
+import type {FeedItem, FeedItemId, XkcdFeedItem} from '@shared/types/feedItems.types';
+import {FeedSourceType} from '@shared/types/feedSources.types';
+import type {FeedSource} from '@shared/types/feedSources.types';
 import {fromQueryFilterOp} from '@shared/types/query.types';
 import type {AsyncResult} from '@shared/types/results.types';
 import type {UserFeedSubscription} from '@shared/types/userFeedSubscriptions.types';
